Link with validator_ragel only in standalone binaries
That should be enough for the testing phase we are at now. Yet another difference is a bit ugly (esp in GYP), I am sure it will go away in months, not years.

This change unblocks NaCl DEPS roll in Chromium caused by commit on revision 8372. I performed the $TEST manually on linux and it passed.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2597
TEST=pull into chrome build, compare chrome binary size, should *not* change dramatically

Review URL: https://chromiumcodereview.appspot.com/10174016

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8401 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2