Add NaClFastMutex which uses pthread mutex and CRITICAL_SECTION locks.
NaClFastMutex is an abstraction which may be implemented later with
spin locks and is not guaranteed to work with condition variables.
This CL modifies the refcount base class to use NaClFastMutexes and
the NaClApp's desc_mu lock to also use NaClFastMutexes, so accessing
descriptors on Windows should be faster.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2741
TEST= existing tests
Review URL: https://chromiumcodereview.appspot.com/10146025

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8403 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2