Clean up duplicated signal number list by using NACL_ARRAY_SIZE
This requires changing some index types to avoid warnings.

Also switch from "#ifdef SIGSTKFLT" to "#if OS_LINUX" because feature
tests are not the usual style in the NaCl codebase, and the latter
makes it clearer to the reader when the conditional will apply.

BUG=none
TEST=build

Review URL: https://chromiumcodereview.appspot.com/10383079

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8530 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2