Remove tests/nacl_file/nacl_file.c, which is not built or tested
This was originally added here with the message:

"Add a copy of nacl_file to the tests directory for experimentation.
This is the old version from naclports photo."

but it is not used, and it refers to a function that no longer exists,
srpc_get_fd().

BUG=none
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10384096

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8558 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2