Remove '-g' from build options for ARM because it breaks the Gyp IRT build
This fixes a build failure on a Chromium ARM bot introduced by my
earlier change, r8671.

This brings the Gyp build into line with the Scons build.

The NaCl build builds the full irt.nexe in Scons but not in Gyp, which
is why I didn't catch this before.  I hadn't realised there was a
Chromium bot that was building the ARM irt.nexe already.

BUG=none
TEST=To test this build with Gyp I used:
  "./native_client/build/gyp_nacl native_client/build/all.gyp ppapi/native_client/native_client.gyp -D disable_nacl=0 -D disable_nacl_untrusted=0 -D target_arch=arm"
  (with various env vars set, copied from NaCl's buildbot scripts)

Review URL: https://chromiumcodereview.appspot.com/10441003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8699 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2