Stop uploading with caching turned off.
In the past because we were not archiving to unique names everywhere,
we developed the practice of turning off caching for GSD uploads to make
read/write atomic.
This likely reduces performance substantially (contributing to flake)
and is now unneeded as we have unique names for almost everything.

Turning this off for all uploads except the ones that appear to need it.

BUG=None
TEST=None
R=mseaborn@chromium.org,noelallen@chromium.org,robertm@chromium.org,khim@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10444099

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8742 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2