Making nacl's default asm rule optional, and using in tools.gyp.
Nacl's default rule to build .S files with a hermetic assembler
was not meant to be turned on for the .S files in tools.gyp which
are meant to be cross-compiled with the nacl toolchains.
~Fortunately~ a limitation of visual studio caused actions to take
precidence over rules, blocking the .S file rule from kicking in
(as the nacl nlib_target variable brought in nacl's cross compile action).
Windows ninja lacks this limitation, thus attempting to build the .S file
with the native compiler in addition to the nacl toolchains.

Adding a flag to allow individual targets to disable the default
asm rule.

BUG=None
TEST=build
R=scottmg@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10477002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8787 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2