Simplify creation and teardown of NaClDescEffectors
Remove the Dtor() method of NaClDescEffector, since the
implementations of it don't do anything apart from unset the vtable
pointer.  NaClDescEffectors have very simple lifetimes -- a function
receiving one never saves a pointer to it -- so a dynamic destructor
is not needed.

Allocate a single global instance of NaClDescEffectorTrustedMem, so
that we don't have to create instances of this dynamically.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2550
TEST=existing tests

Review URL: https://chromiumcodereview.appspot.com/10496014

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8906 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2