Clean up toolchain command/option selection
This cleans up the selection of native-code toolchain details to be
more clear and table-driven.  It also drops the support for the
obsolete style of tool names such as 'nacl-gcc' and 'nacl64-gcc'.
This now works for the existing standard cases, and also for
naclsdk_mode=custom:... specifying a next-generation toolchain
that uses 'arm-nacl-...' or 'i686-nacl-...'.

BUG= none
TEST= trybots

R=bradnelson@google.com

Review URL: https://chromiumcodereview.appspot.com/10704036

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9060 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2