x86-32: Make the gs segment read-only for untrusted code
This is in preparation for using "jmp *%gs:CONST" for jumping to an
address previously vetted by the TCB.

Also, this stops programs from depending on behaviour that we have
said is not allowed by the ABI.  We said some time ago that the only
allowed instruction for accessing the gs segment on x86-32 should be
to read %gs:0 using "movl %gs:0, REG".  We haven't tightened up the
x86-32 validator to prohibit other %gs accesses yet, but making the gs
segment read-only discourages using one class of gs-prefixed
instructions.

This adds a test to check that the gs segment is read-only and also
limited to 4 bytes in size.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2250
BUG=http://code.google.com/p/nativeclient/issues/detail?id=2881
TEST=run_gs_segment_read_only_test etc.

Review URL: https://chromiumcodereview.appspot.com/10702122

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9129 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2