Remove use of the llvm.nacl.thread.stack.padding intrinsic
This is no longer necessary because the thread_create() IRT call (and
the syscall) aligns the stack appropriately on behalf of user code.

We leave in the alignment code for x86 so that new nexes will still
work on older versions of NaCl/Chrome where thread_create() doesn't
align the stack.  There's no need to do this for ARM because there are
no officially released versions of ARM NaCl/Chrome.

Remove __nacl_thread_stack_padding() from tls_params.h because it's
not a TLS parameter, and because it's clearer if the comments about
compatibility can be kept in one place, next to kStackAlignment in
nc_thread.c.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2904
TEST=run_stack_alignment_test with realignment removed from NaClCommonSysThread_Create()

Review URL: https://chromiumcodereview.appspot.com/10827094

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9326 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2