Debug stub: Test single-stepping/continuing with other threads suspended
The debug stub allows GDB to single-step or a continue one thread
while all other threads remain suspended.  This change adds tests to
cover this.  The test checks that other threads remain properly
suspended.

This is in preparation for changing how the debug stub suspends
threads -- the new test catches bugs in my work-in-progress change
which are otherwise not detected.

This test checks whether variables get modified by suspended threads.
We use "nm" to find variables' addresses.  I'm not attempting to make
this work with non-ARM PNaCl.  I've changed the nacl.scons file to use
AllowInlineAssembly() for consistency with other assembly-using tests
where we don't attempt to get x86 inline assembly to work with PNaCl.

We move AssertReplySignal() to the top level so that it can be used by
the new tests, which are in a separate class to provide a new scope
for clarity.  This involves defining an AssertEquals() that is usable
outside of the class.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2948
TEST=run_debug_stub_test

Review URL: https://chromiumcodereview.appspot.com/10829230

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9416 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2